Revert "Feature admin frontend addmovie"
