+++ package.json
@@ -32,8 +32,7 @@
         "react-query": "^3.39.3",
         "react-router-dom": "^6.25.1",
         "react-toastify": "^10.0.6",
-        "typescript": "^5.6.3",
-        "zod": "^3.23.8"
+        "typescript": "^5.6.3"
     },
     "devDependencies": {
         "@eslint/js": "^9.9.0",
