--- src/componets/ProtectedRoute.tsx
@@ -1,18 +1,0 @@
-import { useAuth } from '../context/authContext'
-import { Navigate, Outlet } from 'react-router-dom'
-import AdminLayout from './admin/AdminLayout'
-type Roles = 'admin' | 'user'
-export default function ProtectedRoute({ role }: { role: Roles }) {
-    const { token, user } = useAuth()
-    if ((role && user && user.role !== role) || !token) {
-        return <Navigate to="/" />
-    }
-    if (role === 'admin') {
-        return (
-            <AdminLayout>
-                <Outlet />
-            </AdminLayout>
-        )
-    }
-    return <Outlet />
-}
