+++ src/main.tsx
@@ -6,7 +6,6 @@ import '../sass/main.scss'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 import routes from './utils/routes'
 import App from './App'
-import ProtectedRoute from './componets/ProtectedRoute'
 
 // set a title
