+++ src/utils/api/auth/useSignIn.ts
@@ -11,17 +11,13 @@ export const useSignIn = () => {
         mutationFn: signIn,
         onSuccess: (data) => {
             sessionStorage.setItem('token', JSON.stringify({ token: data.bearer }))
-            sessionStorage.setItem(
-                'user',
-                JSON.stringify({ firstName: data.firstName, lastName: data.lastName, role: data.role })
-            )
+            sessionStorage.setItem('user', JSON.stringify({ firstName: data.firstName, lastName: data.lastName }))
             setTimeout(() => {
                 // tar bort alla tokens efter 1h
                 sessionStorage.removeItem('token')
                 sessionStorage.removeItem('user')
             }, 1000 * 60 * 60)
-            if (data.role === 'admin') window.location.replace('/admin')
-            else window.location.replace('/medlem')
+            window.location.replace('/medlem')
         },
         onError: (error) => {
             console.error('Error signing in', error)
