+++ src/utils/types/types.ts
@@ -29,7 +29,7 @@ export type USERBOOKING = {
     seats: SEAT[]
 }
 
-export type CONFIRMMODALPROPS = {
+export type CONFIRMDELETEMODALPROPS = {
     show: boolean
     onHide: () => void
     onConfirm: () => void
@@ -42,7 +42,6 @@ export type SIGNINRESPONSE = {
     bearer: string
     firstName: string
     lastName: string
-    role: string
 }
 export type SIGNUP = {
     firstName: string
@@ -54,7 +53,7 @@ export type SIGNUPRESPONSE = {
     changes: string
     lastInsertRowid: string
 }
-export type Description = {
+type Description = {
     director: string
     cast: string[]
     synopsis: string
@@ -126,34 +125,4 @@ export type USER = {
     email: string
     firstName: string
     lastName: string
-}
-
-export type NEWMOVIE = {
-    title: string
-    ageLimit: number
-    durationMin: number
-    description: DescriptionNewMovie
-    posterUrl: string
-    trailerUrl: string
-}
-
-export type POSTMOVIERESP = {
-    msg: string
-    movieTitle: string
-    movieId: number
-}
-
-export type MOVIEERROR = {
-    msg: string
-}
-
-export type DescriptionNewMovie = {
-    director: string[]
-    originalTitle: string
-    genre: string[]
-    cast: string[]
-    year: string
-    text: string
-    language: string
-    synopsis: string
 }
